refactor(feishu): extract merge-forward expansion into its own module
Importing parseMergeForwardContent from bot-content.js made send.ts the
third edge of a send -> bot-content -> media -> send cycle that the
topology check rejects. Move the expansion (with its sub-message
formatter) into merge-forward.ts, free of bot-content dependencies;
bot-content re-exports it for existing callers.
